Moving server to a struct
This is some refactoring for the server object. I reduces the number of params that have to be passed to each function by more relying on the config object. Also this allows auto-complete in the IDE so see which methods are available on `server`.

* Move handler logic to its own file
* Make error handling local to server run method

No functional changes were made in this PR.
